Update Debian 10 (buster) installation instruction
Avoid using `apt-key add -` and follow Debian best-practices on how to use 3rd party Debian repository signed keys:

* https://michael-prokop.at/blog/2021/02/16/how-to-properly-use-3rd-party-debian-repository-signing-keys-with-apt/
* https://wiki.debian.org/DebianRepository/UseThirdParty#Sources.list_entry

That way you can also encourage the user to check the GPG fingerprint.

Added Debian codename "buster", so people can find it when searching for it in this repository.

Also removed Bintray URL as it is deprecated and is going "sunset" on May 1st 2021,
see: https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/