[qtwin] Fix crash on pasting screenshots
Workaround for an issue with the Windows Qt version where Celestia crashes
when a copied screenshot is pasted into another application. This appears to
be because the data sent to the clipboard still contains a reference to the
data in the destructed Image object, despite the Qt documentation stating that
QClipboard::setImage causes the clipboard to take ownership of the data. This
does not occur if the QImage manages its own buffer.