QT Celestial Browser - Deep Sky Objects
Within the QT interface is a View -> Celestial Browser -> Deep Sky Objects tab.  It has a filter for the objects consisting of types: Globulars, Galaxies, Nebulae and Open Clusters.  The interface uses check boxes for this filter.  Initially all boxes are checked, but only the Globulars are displayed.  Unchecking the Globulars will display only the Galaxies.  Unchecking the Galaxies will display only the Nebulae.  Unchecking the the Nebulae will display only the Open Clusters.  Unchecking the Open Clusters (no buttons checked) displays only the Open Clusters.  So the filter shows only one type of object, irregardless of the number of check boxes checked. This is evident in the type designation displayed and the object count at the bottom of the window.

This behavior is much closer to the behavior of radio buttons, where one button at a time can be selected and only that item is displayed.  The underlying code for the buttons behavior is close to the radio button.  Rewriting the code to behave as check boxes would affect multiple code areas.  It is simple to modify the interface to display radio buttons.  The present grouping of the buttons into the dsoGroup widget will make them exclusive (default), allowing only one button to be pressed at a time.  This will trigger the toggle signal for two buttons (pressing one and unpressing one) and a clicked signal for the button pair.  Keeping the clicked signal allows the existing code to display objects of only the pressed button type.

This behavior makes sense to the user.  Filtering to show only one object type should be expected under these circumstances.  No reports are known of users requiring multiple items to be displayed simultaneously in this tab.  The modification will not affect the way the data is displayed from the current interface.  It only affects the buttons to make them coincide with the filtered display output.