Disable Sonar warnings about manual new/delete on the Value class
Since this class is essentially a custom variant with data stored in the
padding between the discriminant and the union data, there is no way to avoid
needing to do some level of manual memory management without increasing the
class size unnecessarily. In order to reduce noise, this disables Sonar on the
lines triggering the S4963 and S5025 violations.