Repair guided tour so view of object is at expected distance
Interfaces win32 and GTK have a menu item "Tour Guide"; Qt does not. This item displays a selectable list of celestial objects. In version 1.7 the links do not show the objects. The problem was traced to improper distance settings of the gotoSelection code. The distances were not as requested, allowing the view to be too distant, sometimes so close as to be inside the object, and sometimes a negative distance that interfered with other Celestia navigational methods.

The code reads a guide.cel file that allows distances to be in kilometers, astronomical units or light years. The code had a comment that all distances were converted into light years. Logging results showed the conversion was correctly performed. Other Celestia goto codes all convert distances to kilometers. All the codes feed into one common routine for the actual view movement. I cannot find any light year to kilometer conversion in the tour guide codes. Many of the selected items displayed as if the numbers they were sent were at kilometer distances, despite being sent light year distances.

This code changes the converted distance units to kilometers for the tour guide code. It has been tested on GTK3. It is part of the core Celestia code, so functions the same for the win32 interface. (Presently, the win32 interface and GTK interface give identical results for tour guide selections - all matching errors.) Win32 interface uses fixed size dialog box, GTK is dynamic. This code also increases the size of the win32 dialog window.