Better temperature calculation (#2314)
Replace **TempDiscrepancy** with **Emissivity** and **InternalHeatFlux**
- **Emissivity** (unitless, 0-1) represents how efficient a surface
emits thermal energy. This can be used to simulate greenhouse effect.
- **InternalHeatFlux** (W m^-2) is the amount of heat radiated from the
interior per unit surface area, invariant of external factors.

Not sure how correct I did this thing, hopefully I didn't royally screw
something up somewhere.
Also, currently the value of the Stefan-Boltzmann constant is simply
written into the function as is.

This PR has to be approved at the same time as an update in
CelestiaContent which replaces TempDiscrepancy with these values (PR to
be made)!

---------

Co-authored-by: Andrew Tribick <ajtribick@googlemail.com>