Add a setting to control whether gotoSurface preserve initial orientation or not (#2363)
The current implementation preserves orientation, so after you land you
might be looking up or down, not necessarily looking parallel to the
surface.

adding this for VR/game controller, because in VR/AR you would want the
planet surface to be matching the actual environment when you land, the
landing orientation should be consistent.