Remove randstar.celx
This script does not work - the find method takes an internal ID,
which is either the HIP or encoded TYC number. Since the current
database contains a lot of TYC numbers, using a random number
between 0 and the number of stars will most likely result in a
non-existent star number. There is currently no way in the API to
reliably do what this script is trying to do, without having a
densely-packed set of star IDs.